+++ src/main/java/com/verdantartifice/primalmagick/datagen/theorycrafting/IFinishedWeightFunction.java
@@ -1,0 +1,13 @@
+package com.verdantartifice.primalmagick.datagen.theorycrafting;
+
+import com.google.gson.JsonObject;
+
+public interface IFinishedWeightFunction {
+    void serialize(JsonObject json);
+    
+    default JsonObject getWeightJson() {
+        JsonObject json = new JsonObject();
+        this.serialize(json);
+        return json;
+    }
+}